+++ plover_machine_hid.py
@@ -11,11 +11,6 @@ Most buttons have the same names as in GeminiPR, except for the extra buttons
 which are called X1-X26.
 '''
 
-from copy import copy
-from PyQt5.QtCore import QVariant, pyqtSignal
-from PyQt5.QtGui import QIntValidator
-from PyQt5.QtWidgets import QGroupBox, QHBoxLayout, QVBoxLayout, QCheckBox, QLabel, QLineEdit
-
 from plover.machine.base import ThreadedStenotypeBase
 from plover import log
 from plover.misc import boolean
@@ -177,46 +172,3 @@ class HidMachine(ThreadedStenotypeBase):
             "repeat_interval_ms": (50, int),
         }
 
-class HidOption(QGroupBox):
-
-    valueChanged = pyqtSignal(QVariant)
-
-    def checkbox(self, name, label, tooltip):
-        checkbox = QCheckBox(label)
-        checkbox.setToolTip(tooltip)
-        def on_changed(value):
-            self._value[name] = value
-            self.valueChanged.emit(self._value)
-        checkbox.stateChanged.connect(on_changed)
-        self.setters.append(lambda value: checkbox.setChecked(value[name]))
-        return checkbox
-
-    def number(self, name, label, tooltip, minimum, maximum):
-        edit = QLineEdit("")
-        edit.setToolTip(tooltip)
-        edit.setValidator(QIntValidator(minimum, maximum, edit))
-        def on_changed(value):
-            self._value[name] = int(value)
-            self.valueChanged.emit(self._value)
-        edit.textChanged.connect(on_changed)
-        row = QHBoxLayout()
-        row.addWidget(QLabel(label))
-        row.addWidget(edit)
-        self.setters.append(lambda value: edit.setText(str(value[name])))
-        return row
-
-    def __init__(self):
-        super().__init__()
-        self.setters = []
-        self._value = {}
-        vbox = QVBoxLayout()
-        vbox.addWidget(self.checkbox("first_up_chord_send", "Send chord on first key release", "When the first key in a chord is released, the chord is sent.\nIf the key is pressed and released again, another chord is sent."))
-        vbox.addWidget(self.checkbox("double_tap_repeat", "Double tap to repeat", "Tap and then hold a chord to send it repeatedly."))
-        vbox.addLayout(self.number("repeat_delay_ms", "Repeat delay (ms)", "Delay before chord starts repeating.", 10, 10000))
-        vbox.addLayout(self.number("repeat_interval_ms", "Repeat interval (ms)", "Interval between chord repetitions.", 10, 10000))
-        self.setLayout(vbox)
-
-    def setValue(self, value):
-        self._value = copy(value)
-        for setter in self.setters:
-            setter(value)
