+++ setup.cfg
@@ -11,9 +11,10 @@ install_requires =
   hid>=1.0.4,<2
 py_modules =
   plover_machine_hid
+  hid_option
 
 [options.entry_points]
 plover.machine =
   Plover HID = plover_machine_hid:HidMachine
 plover.gui.qt.machine_option =
-  plover_machine_hid:HidMachine = plover_machine_hid:HidOption
+  plover_machine_hid:HidMachine = hid_option:HidOption
